codeIT/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
numpy>=1.24.0
pydantic>=2.6.4
google-generativeai>=0.8.5

